Log: Change uin32_t to size_t for parameter len

+++ Log.cpp
@@ -45,7 +45,7 @@ typedef void (*LogEntryCreatedFct)(
 			const int line,
 			const int16_t code,
 			const char *msg,
-			const uint32_t len);
+			const size_t len);
 
 static LogEntryCreatedFct pFctLogEntryCreated = NULL;
